Split helpers tests into separate files
